added google connect and disconnect

+++ main.py
@@ -2,7 +2,7 @@ import os
 from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Artists, Projects
+from database_setup import Base, Artists, Projects, User
 from werkzeug.utils import secure_filename
 from flask import send_from_directory
 from flask_uploads import UploadSet, configure_uploads, IMAGES
@@ -10,6 +10,8 @@ from flask import session as login_session
 import random, string
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
+from oauth2client.client import AccessTokenCredentials
+
 import httplib2
 import json
 from flask import make_response
@@ -22,9 +24,9 @@ app = Flask(__name__)
 
 CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
-APPLICATION_NAME = "African Artists App"
+APPLICATION_NAME = "African artists"
 
-engine = create_engine('sqlite:///artistdatabase.db')
+engine = create_engine('sqlite:///artistswithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -39,7 +41,7 @@ def showLogin():
     return render_template('login.html', STATE=state)
 
 
-@app.route('/gconnect/', methods=['POST', 'GET'])
+@app.route('/gconnect', methods=['POST', 'GET'])
 def gconnect():
     # Validate state token
     if request.args.get('state') != login_session['state']:
@@ -62,7 +64,7 @@ def gconnect():
 
     # Check that the access token is valid.
     access_token = credentials.access_token
-    url = ('http://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
            % access_token)
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
@@ -101,7 +103,7 @@ def gconnect():
     login_session['gplus_id'] = gplus_id
 
     # Get user info
-    userinfo_url = "http://www.googleapis.com/oauth2/v1/userinfo"
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
 
@@ -110,6 +112,14 @@ def gconnect():
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
+
+    #see if user exists, if it doesn't make a new one
+
+
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
 
     output = ''
     output += '<h1>Welcome, '
@@ -121,6 +131,30 @@ def gconnect():
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
+
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).first()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
 
 # DISCONNECT - Revoke a current user's token and reset their login_session
 
@@ -136,7 +170,7 @@ def gdisconnect():
     print 'In gdisconnect access token is %s', access_token
     print 'User name is: '
     print login_session['username']
-    url = 'http://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
@@ -155,6 +189,89 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
 
+
+@app.route('/fbconnect', methods=['GET', 'POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.8/me"
+    '''
+        Due to the formatting for the result from the server token exchange we have to
+        split the token first on commas and select the first index which gives us the key : value
+        for the server access token then we split it on colons to pull out the actual token value
+        and replace the remaining quotes with nothing so that it can be used directly in the graph
+        api calls
+    '''
+    token = result.split(',')[0].split(':')[1].replace('"', '')
+
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout
+    login_session['access_token'] = token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
+
+
 @app.route('/')
 def homepage():
     return render_template('home.html')
@@ -163,7 +280,10 @@ def homepage():
 def showArtists():
     artists = session.query(Artists).all()
     output = ''
-    return render_template('artists.html', artists=artists)
+    if 'username' not in login_session:
+        return render_template('public_artists.html', artists=artists)
+    else:
+        return render_template('artists.html', artists=artists)
 
 @app.route('/artists/new/', methods =['GET', 'POST'])
 def newArtist():
@@ -180,9 +300,11 @@ def newArtist():
 
 @app.route('/artists/<int:artist_id>/edit/', methods = ['GET', 'POST'])
 def editArtist(artist_id):
+    editedArtistItem = session.query(Artists).filter_by(id=artist_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    editedArtistItem = session.query(Artists).filter_by(id=artist_id).one()
+    if editedArtistItem.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not allowed to edit this artist. Please create your own artist in order to edit.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             editedArtistItem.name = request.form['name']
@@ -203,9 +325,11 @@ def editArtist(artist_id):
 
 @app.route('/artists/<int:artist_id>/delete/', methods = ['GET', 'POST'])
 def deleteArtist(artist_id):
+    artistToDelete = session.query(Artists).filter_by(id=artist_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    artistToDelete = session.query(Artists).filter_by(id=artist_id).one()
+    if artistToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not allowed to edit this artist. Please create your own artist in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(artistToDelete)
         session.commit()
@@ -217,16 +341,23 @@ def deleteArtist(artist_id):
 @app.route('/artists/<int:artist_id>/projects/')
 @app.route('/artists/<int:artist_id>/')
 def showProjects(artist_id):
-    projects = session.query(Projects).all()
+    artist = session.query(Artists).filter_by(id=artist_id).one()
+    creator = getUserInfo(artist.user_id)
+    projects = session.query(Projects).filter_by(artist_id=artist_id).all()
     output = ''
-    return render_template('projects.html', projects=projects)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('public_projects.html', projects=projects, artist=artist, creator=creator)
+    else:
+        return render_template('projects.html', projects=projects, artist=artist, creator=creator)
 
 @app.route('/artists/<int:artist_id>/projects/new/', methods = ['GET', 'POST'])
 def newProject(artist_id):
+    newProjectItem = Projects(title = request.form['title'], description = request.form['description'],imageUrl = request.form['imageUrl'], artist_id=artist_id)
     if 'username' not in login_session:
         return redirect('/login')
+    if editedArtistItem.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not allowed to add project to this artist. Please create your own artist in order to add project.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
-        newProjectItem = Projects(title = request.form['title'], description = request.form['description'],imageUrl = request.form['imageUrl'], artist_id=artist_id)
         session.add(newProjectItem)
         session.commit()
         flash("New Project created")
@@ -236,9 +367,12 @@ def newProject(artist_id):
 
 @app.route('/artists/<int:artist_id>/projects/<int:project_id>/edit/', methods = ['GET', 'POST'])
 def editProject(artist_id, project_id):
+    editedProjectItem = session.query(Projects).filter_by(id=project_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    editedProjectItem = session.query(Projects).filter_by(id=project_id).one()
+        return redirect('/login')
+    if editedProjectItem.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not allowed to edit this artist. Please create your own artist in order to edit.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['title']:
             editedProjectItem.title = request.form['title']
@@ -255,9 +389,11 @@ def editProject(artist_id, project_id):
 
 @app.route('/artists/<int:artist_id>/projects/<int:project_id>/delete/', methods=['GET', 'POST'])
 def deleteProject(artist_id, project_id):
+    projectToDelete = session.query(Projects).filter_by(id=project_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    projectToDelete = session.query(Projects).filter_by(id=project_id).one()
+    if editedProjectItem.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not allowed to edit this artist. Please create your own artist in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(projectToDelete)
         session.commit()
